refactor(gui): split config.rs into config/{mod,types,persistence,view}.rs
The 901-line primer-gui config.rs grew into three distinct
responsibilities living in one file. Split it into a directory module
with a flat `pub use` re-export façade so every external caller keeps
using `crate::config::Foo` unchanged:

- `types.rs`   — the persisted on-disk shapes (GuiConfig + sub-structs/enums)
- `persistence.rs` — CONFIG_FILENAME, ConfigError, config_path, load, save
- `view.rs`    — frontend-facing read/write DTOs (GuiConfigView/Update + From impls)
- `mod.rs`     — module doc + submodule decls + `pub use *` re-exports + tests

Behaviour-preserving: identical 33-symbol pub surface (verified by diff),
205/205 primer-gui tests pass (unchanged from baseline), clippy clean,
fmt clean. The only code change is an explicit `use std::fs;` in
config/tests.rs — the old flat module's `use std::fs` leaked into the
test module via `use super::*`; the re-export façade doesn't carry
private `use` aliases, so the test now imports it directly.

Also fixes two now-stale `primer-gui/src/config.rs` path references in
CLAUDE.md to point at the new module layout.
