+++ src/crates/primer-gui/src/config/tests.rs
@@ -1,3 +1,5 @@
+use std::fs;
+
 use super::*;
 use tempfile::TempDir;
